pyproject.toml
[project]
name = "b2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.2.6",
    "pillow>=11.2.1",
    "requests>=2.32.3",
    "scikit-image>=0.25.2",
]
